+++ extension/markdown-render.js
@@ -19,7 +19,7 @@ import hljs from "./highlightjs/highlightjs.esm.js"
 import { markedHighlight } from "./vendor/marked-highlight.esm.js"
 import markedExtendedTables from "./vendor/marked-extended-tables.esm.js"
 import markedLinkifyIt from "./vendor/marked-linkify-it.esm.js"
-import { createDirectives, presetDirectiveConfigs } from "./vendor/marked-directive.esm.js"
+import { createDirectives } from "./vendor/marked-directive.esm.js"
 import { urlSchemify } from "./marked-link-scheme.esm.js"
 
 import OptionsStore from "./options/options-storage.js"
@@ -30,6 +30,12 @@ const defaultMarkedOptions = Object.assign({}, marked.getDefaults(), {
   headerPrefix: undefined,
   smartypants: undefined,
 })
+
+const directiveConfig = [
+  { level: "container", marker: ":::" },
+  { level: "block", marker: "::" },
+  // { level: "inline", marker: ":" }
+]
 
 export async function resetMarked(userprefs) {
   marked.setOptions(defaultMarkedOptions)
@@ -76,7 +82,7 @@ export async function resetMarked(userprefs) {
     }
     marked.use(markedMath(mathOptions))
   }
-  marked.use(createDirectives([...presetDirectiveConfigs]))
+  marked.use(createDirectives([...directiveConfig]))
   if (userprefs["emoji-shortcode-enabled"]) {
     const { markedEmoji } = await import("./vendor/marked-emoji.esm.js")
     const { default: emojis } = await import("./data/shortcodes.mjs")